Schedule fonctionne - arret de l'utilisation de kheap

+++ include/n7OS/time.h
@@ -10,7 +10,7 @@
 
 void init_time();
 void handler_en_C_TIMER();
-uint64_t get_time();
+uint32_t get_time();
 char *get_time_string();
 
 #endif // N7OS_TIME_H
